Improve errors handling.

+++ src/bl/messages-controller.ts
@@ -1,4 +1,4 @@
-import { Repository } from "typeorm";
+import { FindManyOptions, Repository } from "typeorm";
 import { Message, MessageType } from "../storage/entities/message";
 import { Db } from "../storage/db";
 import { MessagesProcessor } from "./messages-processor";
@@ -26,6 +26,9 @@ export class MessagesController {
 
   public async update(id: number, message: Message): Promise<Message> {    
     let entityToUpdate: Message = await this._repository.findOne(id);
+    if (!entityToUpdate) {
+      return;
+    }
     entityToUpdate.payload = message.payload;
     entityToUpdate.type = undefined;
     let updatedEntity: Message = await this._repository.save(entityToUpdate);
@@ -39,23 +42,30 @@ export class MessagesController {
     return updatedEntity;
   }
 
-  public async getAll(): Promise<Message[]> {
-      return await this._repository.find();           
+  public async getAll(options: FindManyOptions<Message>): Promise<Message[]> {
+    return await this._repository.find(options);             
   }
 
   public async get(id: number): Promise<Message> {
-      return await this._repository.findOne(id);           
+    return await this._repository.findOne(id);           
   }  
 
   public async delete(id: number): Promise<Message> {
     let entityToRemove: Message = await this._repository.findOne(id); 
+    if (!entityToRemove) {
+      return;
+    }
     return await this._repository.remove(entityToRemove);
   }
   
-  private async updateMessageType(id: number, messageType: MessageType): Promise<Message> {
+  private async updateMessageType(id: number, messageType: MessageType): Promise<void> {
     let entityToUpdate: Message = await this._repository.findOne(id);
+    if (!entityToUpdate) {
+      console.log("Unable to update message type, message not found, message id:", id);
+      return;
+    }
     entityToUpdate.type = messageType;
-    return await this._repository.save(entityToUpdate);
+    await this._repository.save(entityToUpdate);
   }
 
   private dealWithMessageType(message: Message): void {
@@ -66,12 +76,12 @@ export class MessagesController {
       .then((messageType: MessageType) => {
         
         // Update the message type in DB.
-        return this.updateMessageType(message.id, messageType);
+        this.updateMessageType(message.id, messageType);
       })
       .catch((err) => {
         console.error(err);
       });
 
     // Exit the method WITHUOT waiting for completion of the async operation.
-  }
+  }  
 }
